fix(dashboard): repair nested admin routes
